# apis4gifts_swaggerdef.yml
swagger: "2.0"
info:
  description: "This project provides a platform to interact via REST APIs with Oracle Autonomous Transaction Procesing Database (ATP) to create and redeem gifts and crosswords."
  version: "2.1.0"
  title: "APIs 4 Gifts Demo"
  #termsOfService: "http://apismadeeasy.cloud/terms"
  contact:
    email: "carlos.rodriguez.iturria@oracle.com"
  #license:
  #  name: "Apache 2.0"
  #  url: "http://www.apache.org/licenses/LICENSE-2.0.html"
basePath: "/"
host: "localhost:3000"
#basePath: "/nf/gifts"
#host: "apipdev.oracleau.cloud"
tags:
- name: "Gifts"
  description: "GPS and other type of Gifts to interact with via mobile application."
  externalDocs:
    description: "Find out more"
    url: "http://apismadeeasy.cloud"
- name: "Crosswords"
  description: " Crosswords to be Redemption via mobile application."
  externalDocs:
    description: "Find out more"
    url: "http://apismadeeasy.cloud"
schemes:
#- "https"
- "http"
security:
  - X-MyAPI-Key: []
paths:
  /gifts/gps:
    get:
      tags:
      - "Gifts"
      summary: "Finds all available GPS Gifts."
      description: "Gets all GPS Gifts"
      operationId: "getGPSGifts"
      produces: 
        - "application/json"
      parameters:
      - name: "gpsGiftId"
        in: "query"
        description: "Field name used to filter the record to be returned."
        required: false
        type: "string"
        x-example: '12345678ABCDEF'
      responses:
        200:
          description: "Returns the full list of GPSGifts."
          schema:
            properties:
              GPSGifts:
                type: "array"
                items:
                  $ref: "#/definitions/GPSGiftsResponse"
    post:
      tags:
      - "Gifts"
      summary: "Creates new GPS Gifts."
      description: "Returns the just inserted record(s)."
      operationId: "postGpsGifts"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Array of GPS Gift(s) that need to be added."
        required: true
        schema:
            properties:
              GPSGifts:
                type: "array"
                items:
                  $ref: "#/definitions/GPSGiftsRequest"
      responses:
        200:
          description: "Returns the just created record(s)."
          schema:
            properties:
              GPSGifts:
                type: "array"
                items:
                  $ref: '#/definitions/GPSGiftsResponse'

  /gifts/gps/redeem:
    get:
      tags:
      - "Gifts"
      summary: "Finds all available Redemptions for GPS Gifts"
      description: "Gets all Redemptions for GPS Gifts"
      operationId: "getRedemptionGPSGifts"
      produces: 
        - "application/json"
      parameters:
      - name: "gpsGiftId"
        in: "query"
        description: "Field name used to filter the record to be returned."
        required: false
        type: "string"
        x-example: '12345678ABCDEF'
      responses:
        200:
          description: "Returns the full list of GPSGifts."
          schema:
            properties:
              RedemptionGPSGifts:
                type: "array"
                items:
                  $ref: "#/definitions/RedemptionGPSGiftsResponse"
    post:
      tags:
      - "Gifts"
      summary: "Creates new Redemption GPS Gifts."
      description: "Returns the just inserted record(s)."
      operationId: "postRedemptionGpsGifts"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Array of Redemption GPS Gift(s) that need to be added."
        required: true
        schema:
            properties:
              RedemptionGPSGifts:
                type: "array"
                items:
                  $ref: "#/definitions/RedemptionGPSGiftsRequest"
      responses:
        200:
          description: "Returns the just created record(s)."
          schema:
            properties:
              RedemptionGPSGifts:
                type: "array"
                items:
                  $ref: '#/definitions/RedemptionGPSGiftsResponse'


  /gifts/redeem:
    get:
      tags:
      - "Gifts"
      summary: "Finds all Redemption available non-GPS Gifts"
      description: "Gets all Redemption available non-GPS Gifts"
      operationId: "getRedemptionGifts"
      produces: 
        - "application/json"
      parameters:
      - name: "giftId"
        in: "query"
        description: "Field name used to filter the record to be returned."
        required: false
        type: "string"
        x-example: '12345678ABCDEF'
      responses:
        200:
          description: "Returns the full list of Gifts."
          schema:
            properties:
              RedemptionGifts:
                type: "array"
                items:
                  $ref: "#/definitions/RedemptionGiftsResponse"
    post:
      tags:
      - "Gifts"
      summary: "Creates new Redemption non-GPS Gifts."
      description: "Returns the just inserted record(s)."
      operationId: "postRedemptionGifts"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Array of Redemption non-GPS Gift(s) that need to be added."
        required: true
        schema:
            properties:
              RedemptionGifts:
                type: "array"
                items:
                  $ref: "#/definitions/RedemptionGiftsRequest"
      responses:
        200:
          description: "Returns the just created record(s)."
          schema:
            properties:
              RedemptionGifts:
                type: "array"
                items:
                  $ref: '#/definitions/RedemptionGiftsResponse'
                  
  /crosswords/redeem:
    get:
      tags:
      - "Crosswords"
      summary: "Finds all Redemption available Crosswords"
      description: "Gets all Redemption available Crosswords"
      operationId: "getCrosswords"
      produces: 
        - "application/json"
      parameters:
      - name: "crosswordId"
        in: "query"
        description: "Field name used to filter the record to be returned."
        required: false
        type: "string"
        x-example: '12345678ABCDEF'
      responses:
        200:
          description: "Returns the full list of Gifts."
          schema:
            properties:
              RedemptionCrosswords:
                type: "array"
                items:
                  $ref: "#/definitions/RedemptionCrosswordsResponse"
    post:
      tags:
      - "Crosswords"
      summary: "Creates new Redemption Crosswords."
      description: "Returns the just inserted record(s)."
      operationId: "postCrosswords"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Array of Redemption Crosswords that need to be added."
        required: true
        schema:
            properties:
              RedemptionCrosswords:
                type: "array"
                items:
                  $ref: "#/definitions/RedemptionCrosswordsRequest"
      responses:
        200:
          description: "Returns the just created record(s)."
          schema:
            properties:
              RedemptionCrosswords:
                type: "array"
                items:
                  $ref: "#/definitions/RedemptionCrosswordsResponse"


securityDefinitions:
  X-MyAPI-Key:
    type: "apiKey"
    name: "x-api-key"
    in: "header"
definitions:
  RedemptionGiftsRequest:
    type: "object"
    required:
    - "gamesourceid"
    - "gamegource"
    - "datetime"
    - "memberid"
    - "redemptionstatus"
    properties:
      gamegourceid:
        type: "string"
        example: "12345678abcdef"
      gamesource:
        type: "string"
        example: "12345678abcdef"
      datetime:
        type: "string"
        example: "2018-09-03T15:00:00"
      memberid:
        type: "string"
        example: "12345"
      redemptionstatus:
        type: "string"
        example: "Redemption"

  RedemptionGiftsResponse:
    type: "object"
    required:
    - "id"
    - "gamesourceid"
    - "gamesource"
    - "datetime"
    - "memberid"
    - "redemptionstatus"
    properties:
      id:
        type: "string"
        example: "12345678abcdef"
      gamesourceid:
        type: "string"
        example: "12345678abcdef"
      gamesource:
        type: "string"
        example: "12345678abcdef"
      datetime:
        type: "string"
        example: "2018-09-03T15:00:00"
      memberid:
        type: "string"
        example: "12345"
      redemptionstatus:
        type: "string"
        example: "Redemption"
        
  GPSGiftsRequest:
    type: "object"
    required:
    - "gpsgiftid"
    - "latitude"
    - "longtitude"
    - "altitude"
    - "imageurl"
    - "redemptionaction"
    properties:
      gpsgiftid:
        type: "string"
        example: "12345678abcdef"
      latitude:
        type: "number"
        example: "-37.835311"
      longtitude:
        type: "number"
        example: "144.9725336"
      altitude:
        type: "number"
        example: "31"
      imageurl:
        type: "string"
        example: "https://i.4pcdn.org/x/1508011708146.png"
      redemptionaction:
        type: "string"
        example: "Active"

  GPSGiftsResponse:
    type: "object"
    required:
    - "id"
    - "gpsgiftid"
    - "latitude"
    - "longtitude"
    - "altitude"
    - "imageurl"
    - "redemptionaction"
    properties:
      id:
        type: "string"
        example: "12345678abcdef"
      gpsgiftid:
        type: "string"
        example: "12345678abcdef"
      latitude:
        type: "number"
        example: "-37.835311"
      longtitude:
        type: "number"
        example: "144.9725336"
      altitude:
        type: "number"
        example: "31"
      imageurl:
        type: "string"
        example: "https://i.4pcdn.org/x/1508011708146.png"
      redemptionaction:
        type: "string"
        example: "Active"
        
  RedemptionGPSGiftsRequest:
    type: "object"
    required:
    - "deviceid"
    - "gpsgiftid"
    - "datetime"
    - "redeemedstatus"
    properties:
      deviceid:
        type: "string"
        example: "12345678abcdef"
      gpsgiftid:
        type: "string"
        example: "12345678abcdef"
      datetime:
        type: "string"
        example: "2018-09-03T15:10:00"
      redeemedstatus:
        type: "string"
        example: "Redemption"

  RedemptionGPSGiftsResponse:
    type: "object"
    required:
    - "id"
    - "deviceid"
    - "gpsgiftid"
    - "datetime"
    - "redeemedstatus"
    properties:
      id:
        type: "string"
        example: "12345678abcdef"
      deviceid:
        type: "string"
        example: "12345678abcdef"
      gpsgiftid:
        type: "string"
        example: "12345678abcdef"
      datetime:
        type: "string"
        example: "2018-09-03T15:10:00"
      redeemedstatus:
        type: "string"
        example: "Redemption"      
        
  RedemptionCrosswordsRequest:
    type: "object"
    required:
    - "deviceid"
    - "crosswordid"
    - "eventdatetime"
    - "eventtype"
    - "engagementcount"
    - "redemptionstatus"
    properties:
      deviceid:
        type: "string"
        example: "12345678abcdef"
      crosswordid:
        type: "string"
        example: "12345678abcdef"
      eventdatetime:
        type: "string"
        example: "2018-09-03T15:10:00"
      eventtype:
        type: "string"
        example: "start"
      engagementcount:
        type: "number"
        example: 45
      redemptionstatus:
        type: "string"
        example: "Redemption"

  RedemptionCrosswordsResponse:
    type: "object"
    required:
    - "id"
    - "deviceid"
    - "crosswordid"
    - "eventdatetime"
    - "eventtype"
    - "engagementcount"
    - "redemptionstatus"
    properties:
      id:
        type: "string"
        example: "12345678abcdef"
      crosswordid:
        type: "string"
        example: "12345678abcdef"
      eventdatetime:
        type: "string"
        example: "2018-09-03T15:10:00"
      eventtype:
        type: "string"
        example: "start"
      engagementcount:
        type: "number"
        example: 45
      redemptionstatus:
        type: "string"
        example: "Redemption"            
            
externalDocs:
  description: "Find out more about our APIs"
  url: "http://apisMadeEasy.cloud"